refactored code out of markdown component
